simplify mod icon loading
